requirements.txt
altair==5.0.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
decorator==5.1.1
dnspython==2.3.0
exceptiongroup==1.1.1
fastapi==0.96.1
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.6.0
isodate==0.6.1
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mysql-connector-python==8.0.33
mysqlclient==2.2.0
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
plotly==5.15.0
protobuf==3.20.3
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
pymongo==4.3.3
Pympler==1.0.1
pynecone==0.1.33
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.13.0
python-engineio==4.4.1
python-multipart==0.0.5
python-socketio==5.8.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.4.2
rsa==4.9
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.27.0
starlette-admin==0.9.0
streamlit==1.23.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typer==0.4.2
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.20.0
validators==0.20.0
watchdog==2.3.1
websockets==10.4
zipp==3.15.0
